backend/requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.3
django-cors-headers==4.3.1